+++ examples/stage/main.c
@@ -20,11 +20,11 @@
 static gboolean
 on_stage_button_press (ClutterStage *stage, ClutterEvent *event, gpointer data)
 {
-  gint x = 0;
-  gint y = 0;
+  float x = 0;
+  float y = 0;
   clutter_event_get_coords (event, &x, &y);
 
-  g_print ("Stage clicked at (%d, %d)\n", x, y);
+  g_print ("Stage clicked at (%f, %f)\n", x, y);
 
   return TRUE; /* Stop further handling of this event. */
 }
